Moves SoilLayer source and header files to /src and /include
